Code review comments incorporated
